requirements.txt
requests
mysql-connector-python
python-dateutil
